+++ ProjectColab/ProjectColab/Modelo/Equipamento.cs
@@ -11,7 +11,7 @@ namespace ProjectColab.Modelo
         public string laboratorio_nome { get; set; }
         public string laboratorio_id { get; set; }
         public string modelo { get; set; }
-        public decimal quantidade { get; set; }
+        public int quantidade { get; set; }
 
         public Equipamento()
         {
@@ -22,7 +22,7 @@ namespace ProjectColab.Modelo
             this.quantidade = 0;
         }
 
-        public Equipamento(string id, string laboratorio_nome, string laboratorio_id,  string modelo, decimal quantidade)
+        public Equipamento(string id, string laboratorio_nome, string laboratorio_id,  string modelo, int quantidade)
         {
             this.id = id;
             this.laboratorio_nome = laboratorio_nome;
